moves api tests to __tests__ directory
