+++ __tests__/api/api.test.js
@@ -1,6 +1,6 @@
 import { describe, beforeEach, afterEach, it, expect } from '@jest/globals';
 import nock from 'nock';
-import { addData, deleteData, getData } from './api';
+import { addData, deleteData, getData } from '../../api/api';
 import mockUsers from './mocks/users';
 
 describe('api tests', () => {
