+++ __tests__/api/auth.test.js
@@ -1,6 +1,6 @@
 import { describe, beforeEach, afterEach, it, expect } from '@jest/globals';
 import nock from 'nock';
-import { postToAuthTokenUrl } from './auth';
+import { postToAuthTokenUrl } from '../../api/auth';
 
 describe('auth api', () => {
   beforeEach(() => {
